requirements.txt
docopt==0.6.2
tenacity==8.1.0
requests==2.28.1
unidiff==0.7.4
pyyaml==6.0
